requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
groq==0.9.0
supabase==2.4.6
httpx==0.27.0
reportlab==4.2.0
python-dotenv==1.0.1
